Files hierarchy reorganization
