+++ src/components/composite/customAGGrid/custom-aggrid-filters/custom-aggrid-filter.tsx
@@ -8,9 +8,9 @@ import React, { ComponentType, MouseEvent, useMemo, useState } from 'react';
 import { Popover } from '@mui/material';
 import { CustomFilterIcon } from './custom-filter-icon';
 import { CustomAggridAutocompleteFilterParams } from './custom-aggrid-autocomplete-filter';
-import { FilterTextComparators, MuiStyles } from '../../../utils';
-import { CustomAggridFilterParams } from './custom-aggrid-types';
-import { useCustomAggridColumnFilter } from './use-custom-aggrid-column-filter';
+import { FilterTextComparators, MuiStyles } from '../../../../utils';
+import { CustomAggridFilterParams } from '../custom-aggrid-types';
+import { useCustomAggridColumnFilter } from './hooks/use-custom-aggrid-column-filter';
 
 const styles = {
     input: {
