+++ src/components/composite/customAGGrid/custom-aggrid-filters/custom-aggrid-text-filter.tsx
@@ -8,8 +8,8 @@ import React, { useMemo } from 'react';
 import { Grid, IconButton, InputAdornment, TextField } from '@mui/material';
 import { Clear as ClearIcon } from '@mui/icons-material';
 import { useIntl } from 'react-intl';
-import { FilterDataTypes, mergeSx, MuiStyles } from '../../../utils';
-import { DisplayRounding } from './display-rounding';
+import { FilterDataTypes, mergeSx, MuiStyles } from '../../../../utils';
+import { DisplayRounding } from '../display-rounding';
 
 const styles = {
     input: {
@@ -34,7 +34,7 @@ interface CustomAggridTextFilterProps {
     decimalAfterDot: number;
 }
 
-export default function CustomAggridTextFilter({
+export function CustomAggridTextFilter({
     value,
     onChange,
     onClear,
