+++ src/components/composite/customAGGrid/utils/custom-aggrid-header-utils.ts
@@ -5,8 +5,8 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import { CustomHeaderComponent } from './custom-aggrid-header';
-import { CustomAggridFilterParams, CustomColDef } from './custom-aggrid-types';
+import { CustomHeaderComponent } from '../custom-aggrid-header';
+import { CustomAggridFilterParams, CustomColDef } from '../custom-aggrid-types';
 
 export const makeAgGridCustomHeaderColumn = <F extends CustomAggridFilterParams = CustomAggridFilterParams>({
     context,
