+++ src/components/composite/index.ts
@@ -14,4 +14,4 @@ export * from './muiTable';
 export * from './reactQueryBuilder';
 export * from './agGridTable';
 export * from './customAGGrid';
-export * from './report-viewer';
+export * from './report';
