+++ src/components/composite/report/index.ts
@@ -5,16 +5,16 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-export * from './log-table';
+export * from './report-viewer/log-table/log-table';
 export * from './QuickSearch';
 export * from './report-severity';
-export * from './report-viewer';
+export * from './report-viewer/report-viewer';
 export * from './report-log.mapper';
-export * from './report-tree.mapper';
-export * from './use-treeview-scroll';
-export * from './treeview-item';
+export * from './report-treeview/report-tree.mapper';
+export * from './report-treeview/hooks/use-treeview-scroll';
+export * from './report-treeview/treeview-item';
 export * from './report.type';
-export * from './report-viewer-context';
-export * from './virtualized-treeview';
+export * from './report-viewer/context/report-viewer-context';
+export * from './report-treeview/virtualized-treeview';
 export * from './report.styles';
 export * from './report.constant';
